src/third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/v0/fullscreen-element-in-v0-expected.txt

CONSOLE ERROR: line 29: Uncaught TypeError: parent.createShadowRoot is not a function
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: parent.createShadowRoot is not a function
Harness: the test ran to completion.
